[WIP] Implement vPortSuppressTicksAndSleep
This is a work in progress implementation - AFAIK it does not even
compile.

The idea will be to go to low power mode sleep on iddle, and restore the
scheduler tick on wake-up. The "only" problem here is the wake-up
itself, for which we probably need an external source like an RTC.

See https://www.freertos.org/low-power-tickless-rtos.html

Signed-off-by: Lluís M. Campos Martínez <lluis@tastaollet.es>